requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Security-Too==5.0.0
email_validator==1.2.1
bcrypt==3.2.2
flask-restful==0.3.9
flask-cors==3.0.10
redis==4.3.4
celery[redis]==5.2.7
weasyprint==57.2
Flask-Caching==2.0.1